feat: conference common filterbar component
