+++ src/components/modalContainer/index.jsx
@@ -3,12 +3,13 @@ import React, { useState } from "react";
 import { IoChevronDownSharp } from "react-icons/io5";
 import Modal from "../modal";
 import { findCategoryData } from "@/data/modalContainerData";
-import ConferencesDropDown from "./pagesDropDown/ConferencesDropDown";
 import AudienceDropDown from "./pagesDropDown/AudienceDropDown";
 import { fetchPodcastByAllFilter } from "@/redux/features/podcast/action";
 import { fetchNewsletterByAllFilter } from "@/redux/features/newsletter/action";
 import { fetchBlogByAllFilter } from "@/redux/features/blog/action";
 import { fetchYoutubeByAllFilter } from "@/redux/features/youtube/action";
+import AreaDropDown from "./pagesDropDown/AreaDropDown";
+import { fetchConferencesByAllFilter } from "@/redux/features/conference/action";
 
 const ModalContainer = ({
   title,
@@ -43,7 +44,7 @@ const ModalContainer = ({
           obj?.node?.name.toLowerCase().includes(searchTerm.toLowerCase())
         )
       : categoryData?.data;
-
+ 
   return (
     <Modal onClose={closeModal}>
       <div className="absolute top-0 left-0 z-10 md:w-[300px] md:top-10 p-[24px] flex flex-col gap-[16px] self-stretch bg-white border border-indigos-op-300 modalShadow rounded-[4px]">
@@ -70,11 +71,14 @@ const ModalContainer = ({
                   {filteredDropDownData?.map((obj) => (
                     <li key={obj.node.id}>
                       {page === "conferences" ? (
-                        <ConferencesDropDown
+                        <AreaDropDown
                           obj={obj.node}
                           categorySelected={categorySelected}
                           menuTitle={menuTitle}
                           setShowModal={setShowModal}
+                          allFilterFunc={fetchConferencesByAllFilter}
+                          pageState={pageState}
+                          page={page}
                         />
                       ) : page === "podcasts" ? (
                         <AudienceDropDown
