+++ src/data/modalContainerData.js
@@ -9,6 +9,7 @@ import {
   setContinentFilter,
   setTechFilter,
 } from "@/redux/features/conference/conferenceSlice";
+import { setHackathonCityFilter, setHackathonContinentFilter, setHackathonCountryFilter, setHackathonTechFilter } from "@/redux/features/hackathon/hackathonSlice";
 import {
   setNewsletterAudienceFilter,
   setNewsletterLangFilter,
@@ -29,21 +30,25 @@ export const findCategoryData = [
   {
     name: "City",
     toChangeAtt: setCityFilter,
+    toChangeHackathonAtt: setHackathonCityFilter,
     isActiveValue: "citySelected",
   },
   {
     name: "Country",
     toChangeAtt: setCountryFilter,
+    toChangeHackathonAtt: setHackathonCountryFilter,
     isActiveValue: "countrySelected",
   },
   {
     name: "Continent",
     toChangeAtt: setContinentFilter,
+    toChangeHackathonAtt: setHackathonContinentFilter,
     isActiveValue: "continentSelected",
   },
   {
     name: "Technology",
     toChangeAtt: setTechFilter,
+    toChangeHackathonAtt: setHackathonTechFilter,
     isActiveValue: "techSelected",
   },
   {
